Add Repository and Entity clases
Add Client Controller, not functional Postman returning error: System.InvalidOperationException: The relationship from 'AppointmentData.User' to 'User' with foreign key properties {'Email' : string} cannot target the primary key {'Id' : Guid} because it is not compatible. Configure a principal key or a set of compatible foreign key properties for this relationship.